--- src/components/presentation/ProcessNavigation.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-import { ConfirmButton } from 'src/features/processEnd/confirm/components/ConfirmButton';
-
-export const ProcessNavigation = () => <ConfirmButton nodeId={'confirm-button'} />;
